Modifcar ubicacion del contexto en Dominio y estructurar carpetas
