+++ ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.ApplicationCore.DataEjemplo/DataAccessManager/UserDam.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CleanArchitecture.ApplicationCore.InterfacesEjemplo.Data;
-using CleanArchitecture.Dominio;
-using CleanArchitecture.Dominio.EntidadesDatabase.Identity;
+using CleanArchitecture.Dominio.Database;
+using CleanArchitecture.Dominio.Database.Entities.Identity;
 using CleanArchitecture.Shared.Peticiones.Responses.User.Usuarios;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
