+++ ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.ApplicationCore.NegocioEjemplo/Negocio/UsersManager/UserNegocio.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using CleanArchitecture.ApplicationCore.InterfacesEjemplo.Data;
 using CleanArchitecture.ApplicationCore.InterfacesEjemplo.Negocio.UsersManager;
-using CleanArchitecture.Dominio.EntidadesDatabase.Identity;
+using CleanArchitecture.Dominio.Database.Entities.Identity;
 using CleanArchitecture.Dominio.Negocio.UsersDto;
 using CleanArchitecture.Shared.Peticiones.Responses.User.Usuarios;
 using Microsoft.Extensions.Logging;
