+++ ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Dominio/Database/Entities/Identity/Role.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace CleanArchitecture.Dominio.Database.Entities.Identity;
+
+public class Role : IdentityRole<int>
+{
+
+}
