--- ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Dominio/EntidadesDatabase/Identity/User.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace CleanArchitecture.Dominio.EntidadesDatabase.Identity
-{
-    public class User : IdentityUser<int>
-    {
-    }
-}
