--- ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Dominio/EntidadesDatabase/Identity/UserRole.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace CleanArchitecture.Dominio.EntidadesDatabase.Identity
-{
-    public class UserRole : IdentityUserRole<int>
-    {
-
-    }
-}
