--- ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Dominio/EntidadesDatabase/Identity/UserToken.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace CleanArchitecture.Dominio.EntidadesDatabase.Identity
-{
-    public class UserToken : IdentityUserToken<int>
-    {
-
-    }
-}
