+++ ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Ejemplo/Extensions/ServiceCollectionExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using CleanArchitecture.ApplicationCore.DataEjemplo;
 using CleanArchitecture.ApplicationCore.NegocioEjemplo;
-using CleanArchitecture.Dominio;
-using CleanArchitecture.Dominio.EntidadesDatabase.Identity;
+using CleanArchitecture.Dominio.Database;
+using CleanArchitecture.Dominio.Database.Entities.Identity;
 using CleanArchitecture.Infraestructure.DatabaseConfig;
 using Microsoft.AspNetCore.Identity;
