+++ ApuntesNet/ApuntesNet/Arquitectura/CleanArchitecture/CleanArchitecture.Infraestructure.InitDatabase/Program.cs
@@ -1,5 +1,5 @@
-﻿using CleanArchitecture.Dominio;
-using CleanArchitecture.Dominio.EntidadesDatabase.Identity;
+﻿using CleanArchitecture.Dominio.Database;
+using CleanArchitecture.Dominio.Database.Entities.Identity;
 using CleanArchitecture.Infraestructure.InitDatabase;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
